animation for beat

+++ Util.pde
@@ -1,6 +1,6 @@
 public float round(float value, int place){
   float output = (float)(value * Math.pow(10,place));
-  output = Math.round(output);
+  output = (float)(Math.round(output));
   output /= Math.pow(10,place);
   return output;
 }
